+++ CHANGELOG.md
@@ -11,7 +11,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## 2023-05-04: v0.14.1
 ### Fixed
-- Import crypto module conditionally when in Node.js environment. For browsers and bundlers no module import is required and the Web API module is used.
+- Import crypto module conditionally when in Node.js environment. For browsers and bundlers no module import is required and the Web API module is used. Add
+```javascript
+...
+    externals: {
+        "node:crypto": {}
+    },
+...
+```
+to Webpack config.
 
 ## 2023-05-01: v0.14.0
 ### Changed